Module stylist::ast

This module contains the semantic representation of a CSS StyleSheet.

Sheet
└── Vec<enum ScopeContent>
    ├── Block
    │   ├── condition: Vec<Selector>
    │   │   └── fragments: Vec<StringFragment>
    │   └── content: Vec<enum RuleBlockContent>
    │       ├── StyleAttr
    │       │   ├── key: String
    │       │   └── value: Vec<StringFragment>
    │       ├── Block (*)
    │       └── Rule (*)
    └── Rule
        ├── condition: Vec<StringFragment>
        └── Vec<enum RuleBlockContent (*)>

Warning

This module is not stable at the moment and is exposed to be used by procedural macros. Its API may change at anytime.

Macros

sheet

A procedural macro that parses a string literal or an inline stylesheet into a Sheet.

Structs

Block

A block is a set of css properties that apply to elements that match the condition. The CSS standard calls these “Qualified rules”.

Rule

An At-Rule can contain both other blocks and in some cases more At-Rules.

Selector

A CSS Selector.

Sheet

The top node of a stylesheet.

StringFragment

A String Fragment

StyleAttribute

A simple CSS property in the form of a key value pair. Mirrors what would be called a “Declaration” in the CSS standard.

StyleContext

A context to faciliate ToStyleStr.

Enums

Bow

A clone-on-write smart pointer with boxed owned data.

RuleBlockContent

The content of a Rule or a Block

ScopeContent

A scope represents a media query or all content not in a media query. The CSS-Syntax-Level-3 standard calls all of these rules, which is used here specifically for At-Rules. A Qualified rule is represented by a Block, an At-Rule is represented by a Rule.

Traits

ToStyleStr

Structs implementing this trait should be able to turn into a part of a CSS style sheet.