Macro stylist::css

css!() { /* proc-macro */ }

A procedural macro that parses a string literal or an inline stylesheet into a StyleSource.

Please consult the documentation of the macros module for the supported syntax of this macro.

Example

use stylist::css;
use stylist::yew::Global;
use yew::prelude::*;

let rendered = html! {<div class={css!("color: red;")} />};
let rendered_global = html! {<Global css={css!("color: red;")} />};