Module stylist::manager

Customise behaviour of Styles.

This module contains StyleManager which can be used for customising mounting point / mounting behaviour for styles (when rendering contents into a ShadowRoot or an <iframe />).

This is an advanced feature and most of the time you don’t need to use it.

Structs

StyleId

The Unique Identifier of a Style.

StyleManager

A struct to customise behaviour of Style.

StyleManagerBuilder

A builder for StyleManager.