Struct stylist::Style

pub struct Style { /* private fields */ }

A struct that represents a scoped Style.

Style Scoping and Substitution Rule for Current Selector(&):

Currently, Stylist processes selectors with the following rules:

If a style attribute is not in any block (dangling style attribute), it will be scoped with the generated class name. (Applied to the element where the generated class name is applied to.)

For example, for the following style:

color: red;

Stylist will generate the following stylesheet:

.stylist-uSu9NZZu {
    color: red;
}

For style attributes that are in a block, for each selector of that block, the following rules apply:

If a selector contains a Current Selector(&), the current selector will be substituted with the generated class name.

If a selector starts with a pseudo-class selector, it will be applied to the root element.

For other selectors, it will be prefixed with the generated class name.

Example, original style:

&.invalid input, input:invalid {
    color: red;
}

:hover {
    background-color: pink;
}

.hint {
    font-size: 0.9rem;
}

Stylist generated stylesheet:

.stylist-uSu9NZZu.invalid input, .stylist-uSu9NZZu input:invalid {
    color: red;
}

.stylist-uSu9NZZu:hover {
    background-color: pink;
}

.stylist-uSu9NZZu .hint {
    font-size: 0.9rem;
}

Note:

Root pseudo class (:root) will also be treated like a Current Selector.

Implementations

impl Style

pub fn new<Css>(css: Css) -> Result<Self>where
Css: TryInto<StyleSource>,
Error: From<Css::Error>,

Creates a new style from some parsable css with a default prefix.

Examples

use stylist::Style;

let style = Style::new("background-color: red;")?;

pub fn create<N, Css>(class_prefix: N, css: Css) -> Result<Self>where
N: Into<Cow<'static, str>>,
Css: TryInto<StyleSource>,
Error: From<Css::Error>,

Creates a new style with a custom class prefix from some parsable css.

Examples

use stylist::Style;

let style = Style::create("my-component", "background-color: red;")?;

pub fn new_with_manager<Css, M>(css: Css, manager: M) -> Result<Self>where
Css: TryInto<StyleSource>,
Error: From<Css::Error>,
M: Into<StyleManager>,

Creates a new style from some parsable css with a default prefix using a custom manager.

pub fn create_with_manager<N, Css, M>(
class_prefix: N,
css: Css,
manager: M
) -> Result<Self>where
N: Into<Cow<'static, str>>,
Css: TryInto<StyleSource>,
Error: From<Css::Error>,
M: Into<StyleManager>,

Creates a new style with a custom class prefix from some parsable css using a custom manager.

pub fn get_class_name(&self) -> &str

Returns the class name for current style

You can add this class name to the element to apply the style.

Examples

use stylist::Style;

let style = Style::create("my-component", "background-color: red;")?;

// Example Output: my-component-uSu9NZZu
println!("{}", style.get_class_name());

pub fn get_style_str(&self) -> &str

Get the parsed and generated style in &str.

This is usually used for debug purposes or testing in non-wasm32 targets.

Examples

use stylist::Style;

let style = Style::create("my-component", "background-color: red;")?;

// Example Output:
// .my-component-uSu9NZZu {
//     background-color: red;
// }
println!("{}", style.get_style_str());

pub fn unregister(&self)

Unregister current style from style registry.

After calling this method, the style will be unmounted from DOM after all its clones are freed.

Note

Most of time, you don’t need to unmount a style.

pub fn id(&self) -> &StyleId

Returns the StyleId for current style.