Struct stylist::StyleSource

pub struct StyleSource { /* private fields */ }

A struct that can be used as a source to create a Style or GlobalStyle.

This struct is usually created by css! macro.

You can also get a StyleSource instance from a string or a Sheet by calling .into().

use stylist::yew::Global;
use stylist::{css, StyleSource};
use yew::prelude::*;

let s: StyleSource = css!("color: red;");

let rendered = html! {<div class={s.clone()} />};
let global_rendered = html! {<Global css={s} />};