Module stylist::yew

This module contains yew specific features.

Usage in function components

You can create a style and use it like this:

use stylist::yew::use_style;
use yew::prelude::*;

#[function_component]
fn MyStyledComponent() -> Html {
    let style = use_style!("color: red;");
    html! {<div class={style}>{"Hello World!"}</div>}
}

Structs

Global

A Global Style that will be applied to <html /> tag, inspired by emotion.

GlobalProps

The properties for the Global Component, please see its documentation for usage.

ManagerProvider

A Context Provider to provide a custom StyleManager to child components.

ManagerProviderProps

The properties for ManagerProvider Component, please see its documentation for usage.

Attribute Macros

styled_component

A procedural macro to style a function component.

styled_component_impl

A procedural macro to use a specialized, contextual css! macro.