Attribute Macro stylist::yew::styled_component_impl

#[styled_component_impl]

A procedural macro to use a specialized, contextual css! macro.

styled_component is implemented in terms of this, prefer that if possible. If you need to use function_component directly but still inject the contextual css! macro, use this.

You can also use the attribute on functions that have access to Hooks to enable the usage of a contextual css! in their body.

Example:

use stylist::yew::styled_component_impl;
use yew::prelude::*;

// Equivalent to #[styled_component(MyStyledComponent)]
// This usage is discouraged, prefer `styled_component`
#[styled_component_impl]
#[function_component(MyStyledComponent)]
fn my_styled_component() -> Html {
    html! {<div class={css!("color: red;")}>{"Hello World!"}</div>}
}