Struct stylist::yew::Global

pub struct Global { /* private fields */ }

A Global Style that will be applied to <html /> tag, inspired by emotion.

The css attribute accepts a value of any type that implements IntoPropValue<StyleSource>. If you use the parser feature, this panics when supplying a string that fails to parse.

The style will be applied via the :root pseudo-class, which has higher specificity than “simply” using the html selector.

Example:

use stylist::css;
use stylist::yew::Global;
use yew::prelude::*;

#[function_component]
fn App() -> Html {
    html! {
        <>
            <Global css={css!("color: red;")} />
            <div>{"Hello World!"}</div>
        </>
    }
}