Attribute Macro stylist::yew::styled_component

#[styled_component]

A procedural macro to style a function component.

Specifically this introduces a specialized css! macro that is aware of the contextual style manager.

For detailed arguments and usage see also the underlying function_component attribute in Yew.

Example:

use std::borrow::Cow;

use stylist::yew::styled_component;
use yew::prelude::*;

#[styled_component]
fn MyStyledComponent() -> Html {
    html! {<div class={css!("color: red;")}>{"Hello World!"}</div>}
}

Note:

You don’t need to import css! inside of a styled_component.